refactor(www/audio): revert to direct download instead of downloading a Blob directly
