+++ apps/www/components/Audio/AudioPlayer/ExpandedAudioPlayer.tsx
@@ -18,10 +18,10 @@ import Queue from './ui/tabs/queue/Queue'
 import AudioControl, { AudioControlProps } from './controls/AudioControl'
 import LatestArticles from './ui/tabs/latest/LatestArticles'
 import { AudioQueueItem } from '../graphql/AudioQueueHooks'
-import { downloadFileFromUrl } from '../../../lib/helpers/FileDownloadHelper'
 import AudioError from './ui/AudioError'
 import { NEXT_PUBLIC_FEAT_HOERT_HOERT } from '../constants'
 import { useUserAgent } from '../../../lib/context/UserAgentContext'
+import downloadAudioSourceFile from '../helpers/DownloadAudioSource'
 
 const styles = {
   root: css({
@@ -161,16 +161,7 @@ const ExpandedAudioPlayer = ({
 
   const handleDownload = async (item: AudioQueueItem['document']) => {
     try {
-      const {
-        meta: { audioSource, title },
-      } = item
-      const downloadSource =
-        audioSource.mp3 || audioSource.aac || audioSource.ogg
-      const extension = downloadSource.split('.').pop()
-      const serializedTitle = title.replace(/[^a-z0-9]/gi, '_').toLowerCase()
-      const filename = `${serializedTitle}-republik.${extension}`
-
-      await downloadFileFromUrl(downloadSource, filename)
+      downloadAudioSourceFile(item)
     } catch (err) {
       // TODO: handle download error
       console.error(err)
